+++ OData2PocoLib/Api/O2P.cs
@@ -13,10 +13,12 @@ namespace OData2Poco.Api
         internal MetaDataInfo MetaData { get; set; }
         public string MetaDataAsString => MetaData?.MetaDataAsString;
         public string MetaDataVersion => MetaData?.MetaDataVersion;
-        public string SchemaNamespace =>MetaData?.SchemaNamespace;
+        public string SchemaNamespace => MetaData?.SchemaNamespace;
         public Dictionary<string, string> ServiceHeader => MetaData?.ServiceHeader;
 
         public string CodeText { get; set; }
+        //warning due to renaming of reserved keywords
+        public List<string> ModelWarning => ModelManager.ModelWarning;
         public O2P(PocoSetting setting = null)
         {
             Setting = setting ?? new PocoSetting();
@@ -24,37 +26,23 @@ namespace OData2Poco.Api
             MetaData = new MetaDataInfo();
         }
 
+
+        internal async Task<IPocoGenerator> GenerateModel(OdataConnectionString odataConnString)
+        {
+            IPocoGenerator gen = await PocoFactory.GenerateModel(odataConnString, Setting);
+            MetaData=gen.MetaData;
+            return gen;
+        }
        
-        //internal async Task<IPocoGenerator> GenerateModel(OdataConnectionString odataConnString)
-        //{
-        //    //MetaData = await MetaDataReader.LoadMetadataAsync(odataConnString);
-        //    IPocoGenerator gen = await PocoFactory.GenerateModel(odataConnString, Setting);
-        //    return gen;
-        //}
         public async Task<string> GenerateAsync(OdataConnectionString odataConnString)
         {
-            
-            var generator = await PocoFactory.GeneratePoco(odataConnString, Setting);
-            ClassList = generator.ClassList;
-            CodeText = generator.ToString();
+            var gen = await GenerateModel(odataConnString);
+            var generatorCs = new PocoClassGeneratorCs(gen, Setting);
+            ClassList = generatorCs.ClassList;
+            CodeText = generatorCs.ToString();
             return CodeText;
         }
-        //public async Task<string> GenerateAsync(OdataConnectionString odataConnString)
-        //{
-        //    var gen = await GenerateModel(odataConnString);
-        //    var generatorCs = new PocoClassGeneratorCs(gen, Setting);
-        //    ClassList = generatorCs.ClassList;
-        //    CodeText = generatorCs.ToString();
-        //    return CodeText;
-        //}
-        //public async Task<string> GenerateTsAsync(OdataConnectionString odataConnString)
-        //{
-        //    var gen = await GenerateModel(odataConnString);
-        //    var ts = new PocoClassGeneratorTs(gen, Setting);
-        //    ClassList = ts.ClassList;
-        //    CodeText = ts.ToString();
-        //    return CodeText;
-        //}
+      
 
         public string GenerateProject()
         {